pyproject.toml
[project]
name = "arwi-workbooks"
version = "0.1.0"
description = "Arwi Workbooks API service"
authors = [
    {name = "Kirill Chirkunov",email = "kirill.chirkunov@gmail.com"}
]
license = {text = "Apache License"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "alembic (==1.16.1)",
    "sqlalchemy (>=2.0.32,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "requests (>=2.32.3,<3.0.0)",
    "cachetools (>=5.5.0,<6.0.0)",
    "greenlet (>=3.2.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",

]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
# pycodestyle, pyflakes, isort, pylint, pyupgrade
ignore = ["E501"]
select = ["E", "F", "I", "PL", "UP", "W"]
